move stream title update functionality from the ModView chat dropdown to the kebab menu ⋮
