+++ js/app/components/mobile/desktop-ui/kebab.tsx
@@ -10,10 +10,14 @@ import {
   DropdownMenuTrigger,
   ResponsiveDropdownMenuContent,
   Text,
+  UpdateStreamTitleDialog,
+  useCanModerate,
+  useLivestream,
   useLivestreamInfo,
   useLivestreamStore,
   usePlayerStore,
   useTheme,
+  useUpdateLivestreamRecord,
 } from "@streamplace/components";
 import { EllipsisVertical } from "lucide-react-native";
 import { useState } from "react";
@@ -35,11 +39,20 @@ interface KebabMenuProps {
 export function KebabMenu({ dropdownPortalContainer }: KebabMenuProps) {
   const th = useTheme();
   const [isOpen, setIsOpen] = useState(false);
+  const [showUpdateTitleDialog, setShowUpdateTitleDialog] = useState(false);
 
-  const livestream = useLivestreamStore((x) => x.livestream);
+  const livestreamFromStore = useLivestreamStore((x) => x.livestream);
+  const livestream = useLivestream();
   const setReportModalOpen = usePlayerStore((x) => x.setReportModalOpen);
   const setReportSubject = usePlayerStore((x) => x.setReportSubject);
   const { profile } = useLivestreamInfo();
+
+  // Get the streamer's DID from the profile
+  const streamerDID = profile?.did;
+  // Check moderation permissions for the current user on this streamer's channel
+  const modPermissions = useCanModerate(streamerDID);
+  const { updateLivestream, isLoading: isUpdateTitleLoading } =
+    useUpdateLivestreamRecord();
 
   const iconRotate = useAnimatedStyle(() => {
     return {
@@ -55,31 +68,52 @@ export function KebabMenu({ dropdownPortalContainer }: KebabMenuProps) {
   });
 
   return (
-    <DropdownMenu onOpenChange={setIsOpen} key={dropdownPortalContainer}>
-      <DropdownMenuTrigger>
-        <Animated.View style={[iconRotate]}>
-          <EllipsisVertical color={th.theme.colors.foreground} />
-        </Animated.View>
-      </DropdownMenuTrigger>
-      <ResponsiveDropdownMenuContent
-        side="top"
-        align="end"
-        portalHost={dropdownPortalContainer}
-      >
-        <DropdownMenuGroup title="Report">
-          <ReportStreamItem
-            livestream={livestream}
-            setReportModalOpen={setReportModalOpen}
-            setReportSubject={setReportSubject}
-          />
-          <ReportUserItem
-            profile={profile}
-            setReportModalOpen={setReportModalOpen}
-            setReportSubject={setReportSubject}
-          />
-        </DropdownMenuGroup>
-      </ResponsiveDropdownMenuContent>
-    </DropdownMenu>
+    <>
+      <DropdownMenu onOpenChange={setIsOpen} key={dropdownPortalContainer}>
+        <DropdownMenuTrigger>
+          <Animated.View style={[iconRotate]}>
+            <EllipsisVertical color={th.theme.colors.foreground} />
+          </Animated.View>
+        </DropdownMenuTrigger>
+        <ResponsiveDropdownMenuContent
+          side="top"
+          align="end"
+          portalHost={dropdownPortalContainer}
+        >
+          {modPermissions.canManageLivestream && (
+            <DropdownMenuGroup title="Stream Settings">
+              <UpdateStreamTitleItem
+                setShowUpdateTitleDialog={setShowUpdateTitleDialog}
+                isUpdateTitleLoading={isUpdateTitleLoading}
+                livestream={livestream}
+              />
+            </DropdownMenuGroup>
+          )}
+          <DropdownMenuGroup title="Report">
+            <ReportStreamItem
+              livestream={livestreamFromStore}
+              setReportModalOpen={setReportModalOpen}
+              setReportSubject={setReportSubject}
+            />
+            <ReportUserItem
+              profile={profile}
+              setReportModalOpen={setReportModalOpen}
+              setReportSubject={setReportSubject}
+            />
+          </DropdownMenuGroup>
+        </ResponsiveDropdownMenuContent>
+      </DropdownMenu>
+
+      {showUpdateTitleDialog && (
+        <UpdateStreamTitleDialog
+          livestream={livestream}
+          streamerDID={streamerDID}
+          updateLivestream={updateLivestream}
+          isLoading={isUpdateTitleLoading}
+          onClose={() => setShowUpdateTitleDialog(false)}
+        />
+      )}
+    </>
   );
 }
 
@@ -138,6 +172,32 @@ function ReportUserItem({
       disabled={!profile?.did}
     >
       <Text>Report User...</Text>
+    </DropdownMenuItem>
+  );
+}
+
+function UpdateStreamTitleItem({
+  setShowUpdateTitleDialog,
+  isUpdateTitleLoading,
+  livestream,
+}: {
+  setShowUpdateTitleDialog: (show: boolean) => void;
+  isUpdateTitleLoading: boolean;
+  livestream: any;
+}) {
+  const { onOpenChange } = useRootContext();
+
+  return (
+    <DropdownMenuItem
+      onPress={() => {
+        onOpenChange?.(false);
+        setShowUpdateTitleDialog(true);
+      }}
+      disabled={isUpdateTitleLoading || !livestream}
+    >
+      <Text>
+        {isUpdateTitleLoading ? "Updating..." : "Update stream title"}
+      </Text>
     </DropdownMenuItem>
   );
 }
