+++ js/components/src/index.tsx
@@ -34,6 +34,7 @@ export * from "./lib/theme";
 export * from "./components/chat/chat";
 export * from "./components/chat/chat-box";
 export * from "./components/chat/system-message";
+export * from "./components/chat/update-stream-title-dialog";
 export { default as VideoRetry } from "./components/mobile-player/video-retry";
 export * from "./lib/system-messages";
